+++ CHANGES.rst
@@ -10,6 +10,9 @@ Changelog
 * Use ``display_name`` instead of ``name_get()`` for Odoo >= 8,
   because ``name_get`` is deprecated in Odoo 17 and removed in Odoo 18.
 
+* Remove workaround not needed anymore for interactive console.
+  Issue was fixed in Python source code.
+
 
 2.4.4 (2025-10-16)
 ~~~~~~~~~~~~~~~~~~